move: CreateTeamModal to app/env/[envId]
