+++ apps/web/app/(app)/environments/[environmentId]/settings/members/EditMemberships.tsx
@@ -4,7 +4,7 @@ import ShareInviteModal from "@/app/(app)/environments/[environmentId]/settings/
 import TransferOwnershipModal from "@/app/(app)/environments/[environmentId]/settings/members/TransferOwnershipModal";
 import DeleteDialog from "@/components/shared/DeleteDialog";
 import LoadingSpinner from "@/components/shared/LoadingSpinner";
-import CreateTeamModal from "@/components/team/CreateTeamModal";
+import CreateTeamModal from "@/app/(app)/environments/[environmentId]/CreateTeamModal";
 import { env } from "@/env.mjs";
 import {
   addMember,
